--- tests/test_reasoning_agent_streaming_tool_results.py
@@ -1,526 +1,0 @@
-"""
-Tests for the streaming tool results bug fix.
-
-This module specifically tests that tool results are properly included
-in both streaming and non-streaming reasoning responses.
-
-Background: There was a bug where streaming responses would report tool
-failures even when tools executed successfully, because the final response
-generation wasn't receiving tool results context.
-"""
-
-import os
-from fastmcp import Client, FastMCP
-import pytest
-import json
-import httpx
-from unittest.mock import AsyncMock
-from fastapi.testclient import TestClient
-from api.mcp import to_tools
-from api.reasoning_agent import ReasoningAgent
-from api.models import ChatCompletionRequest, ChatMessage
-from api.tools import ToolResult
-from api.main import app
-from api.dependencies import get_reasoning_agent, get_prompt_manager
-from api.prompt_manager import PromptManager
-from dotenv import load_dotenv
-
-load_dotenv()
-
-
-class TestStreamingToolResultsBugFix:
-    """
-    Test that the streaming tool results bug is fixed.
-
-    THE BUG: Streaming responses would report tool failures even when tools executed
-    successfully, because the final response generation (_stream_final_response) wasn't
-    receiving tool results context from the reasoning process.
-
-    ROOT CAUSE: In the streaming workflow, the reasoning context (which contains tool
-    results) was being stored in self._current_reasoning_context during
-    _stream_reasoning_process, but _stream_final_response wasn't using this context
-    to build reasoning summaries that include tool results.
-
-    SYMPTOMS:
-    - Non-streaming responses correctly included tool results in final synthesis
-    - Streaming responses would say tools "failed" or "didn't execute successfully"
-      even when they actually succeeded
-    - Tool execution events in the stream showed successful results, but the final
-      response ignored them
-
-    THE FIX: Modified _stream_final_response to:
-    1. Use the stored reasoning context (self._current_reasoning_context)
-    2. Call _build_reasoning_summary(reasoning_context) to include tool results
-    3. Add the reasoning summary to the synthesis messages, just like non-streaming
-
-    This ensures streaming and non-streaming responses are consistent and both include
-    actual tool results in their final responses.
-    """
-
-    @pytest.fixture
-    def mock_reasoning_agent(self):
-        """Create a mock reasoning agent for testing."""
-        # Create a real httpx client (required by OpenAI client)
-        http_client = httpx.AsyncClient()
-
-        # Create mocks for MCP and prompt managers
-        tools = []
-        mock_prompt_manager = AsyncMock()
-
-        return ReasoningAgent(
-            base_url="http://test",
-            api_key="test-key",
-            http_client=http_client,
-            tools=tools,
-            prompt_manager=mock_prompt_manager,
-        )
-
-    @pytest.fixture
-    def sample_request(self):
-        """Sample chat completion request."""
-        return ChatCompletionRequest(
-            model="gpt-4o",
-            messages=[
-                ChatMessage(role="user", content="What's the weather in Tokyo?"),
-            ],
-        )
-
-    @pytest.fixture
-    def sample_tool_result(self):
-        """Sample successful tool result."""
-        return ToolResult(
-            tool_name="get_weather",
-            success=True,
-            result={
-                "location": "Tokyo",
-                "temperature": "22°C",
-                "condition": "Sunny",
-            },
-            execution_time_ms=150.0,
-        )
-
-    @pytest.mark.asyncio
-    async def test_stream_final_response_includes_tool_results(
-        self, mock_reasoning_agent: ReasoningAgent, sample_request: ChatCompletionRequest, sample_tool_result: ToolResult,  # noqa: E501
-    ):
-        """Test that _stream_final_response would include tool results by verifying message structure."""  # noqa: E501
-        # Setup mocks
-        mock_reasoning_agent.prompt_manager.get_prompt.return_value = "You are a helpful assistant."  # noqa: E501
-
-        # Create reasoning context with tool results (the bug was that this wasn't passed to streaming)  # noqa: E501
-        reasoning_context = {
-            "steps": [],
-            "tool_results": [sample_tool_result],
-            "final_thoughts": "",
-            "user_request": sample_request,
-        }
-
-        # Test the internal message building logic that was part of the bug fix
-        synthesis_prompt = await mock_reasoning_agent.prompt_manager.get_prompt("final_answer")
-
-        # Build synthesis messages (this is what the bug fix added to streaming)
-        messages = [
-            {"role": "system", "content": synthesis_prompt},
-            {"role": "user", "content": f"Original request: {sample_request.messages[-1].content}"},  # noqa: E501
-        ]
-
-        # Add reasoning summary if available (this was missing in streaming before fix)
-        if reasoning_context:
-            reasoning_summary = mock_reasoning_agent._build_reasoning_summary(reasoning_context)
-            messages.append({
-                "role": "assistant",
-                "content": f"My reasoning process:\n{reasoning_summary}",
-            })
-
-        # Verify the fix: streaming now includes tool results like non-streaming
-        assert len(messages) == 3  # system + user + assistant (reasoning summary)
-        assert messages[0]["role"] == "system"
-        assert messages[1]["role"] == "user"
-        assert messages[2]["role"] == "assistant"
-
-        # Verify that tool results are included in the reasoning summary
-        reasoning_message = messages[2]["content"]
-        assert "Tool Results:" in reasoning_message
-        assert "get_weather" in reasoning_message
-        assert "Tokyo" in reasoning_message
-        assert "22°C" in reasoning_message
-
-    @pytest.mark.asyncio
-    async def test_stream_final_response_without_reasoning_context(
-        self, mock_reasoning_agent: ReasoningAgent, sample_request: ChatCompletionRequest,
-    ):
-        """Test that _stream_final_response handles missing reasoning context gracefully."""
-        # Setup mocks
-        mock_reasoning_agent.prompt_manager.get_prompt.return_value = "You are a helpful assistant."  # noqa: E501
-
-        # Test the internal message building logic with None context
-        synthesis_prompt = await mock_reasoning_agent.prompt_manager.get_prompt("final_answer")
-
-        # Build synthesis messages without reasoning context
-        messages = [
-            {"role": "system", "content": synthesis_prompt},
-            {"role": "user", "content": f"Original request: {sample_request.messages[-1].content}"},  # noqa: E501
-        ]
-
-        # Add reasoning summary if available (None in this case)
-        reasoning_context = None
-        if reasoning_context:
-            reasoning_summary = mock_reasoning_agent._build_reasoning_summary(reasoning_context)
-            messages.append({
-                "role": "assistant",
-                "content": f"My reasoning process:\n{reasoning_summary}",
-            })
-
-        # Verify that it handles None gracefully - no reasoning summary added
-        assert len(messages) == 2  # system + user only (no reasoning summary)
-        assert messages[0]["role"] == "system"
-        assert messages[1]["role"] == "user"
-
-    @pytest.mark.asyncio
-    async def test_build_reasoning_summary_includes_tool_results(
-        self, mock_reasoning_agent: ReasoningAgent, sample_tool_result: ToolResult,
-    ):
-        """Test that _build_reasoning_summary properly includes tool results."""
-        reasoning_context = {
-            "steps": [],
-            "tool_results": [sample_tool_result],
-            "final_thoughts": "",
-        }
-
-        summary = mock_reasoning_agent._build_reasoning_summary(reasoning_context)
-
-        # Verify tool results are included in summary
-        assert "Tool Results:" in summary
-        assert "get_weather" in summary
-        assert "Tokyo" in summary
-        assert "22°C" in summary
-
-    @pytest.mark.asyncio
-    async def test_build_reasoning_summary_with_multiple_tool_results(
-        self, mock_reasoning_agent: ReasoningAgent,
-    ):
-        """Test that _build_reasoning_summary handles multiple tool results."""
-        tool_results = [
-            ToolResult(
-                tool_name="get_weather",
-                success=True,
-                result={"location": "Tokyo", "temp": "22°C"},
-                execution_time_ms=100.0,
-            ),
-            ToolResult(
-                tool_name="search_web",
-                success=True,
-                result={"query": "weather", "results": ["result1"]},
-                execution_time_ms=200.0,
-            ),
-        ]
-
-        reasoning_context = {
-            "steps": [],
-            "tool_results": tool_results,
-            "final_thoughts": "",
-        }
-
-        summary = mock_reasoning_agent._build_reasoning_summary(reasoning_context)
-
-        # Verify both tool results are included
-        assert "get_weather" in summary
-        assert "search_web" in summary
-        assert "Tokyo" in summary
-        assert "weather" in summary
-
-    @pytest.mark.asyncio
-    async def test_current_reasoning_context_is_stored(
-        self, mock_reasoning_agent: ReasoningAgent, sample_request: ChatCompletionRequest,
-    ):
-        """Test that reasoning context is stored in the instance for streaming access."""
-        mock_reasoning_agent.prompt_manager.get_prompt.return_value = "system prompt"
-        # Tool availability is determined by the tools list in the constructor
-
-        # Mock the OpenAI client response
-        mock_openai_response = AsyncMock()
-        mock_openai_response.choices = [AsyncMock()]
-        mock_openai_response.choices[0].message.content = '{"thought": "I need to help the user", "next_action": "FINISHED", "tools_to_use": [], "concurrent_execution": false}'  # noqa: E501
-
-        # Create a proper mock for the openai client
-        mock_reasoning_agent.openai_client = AsyncMock()
-        mock_reasoning_agent.openai_client.chat.completions.create.return_value = mock_openai_response  # noqa: E501
-
-        # Call the streaming reasoning process and consume all events
-        chunks = []
-        generator = mock_reasoning_agent._stream_reasoning_process(
-            sample_request, "test-id", 1234567890,
-        )
-
-        try:
-            async for chunk in generator:
-                chunks.append(chunk)
-                # Continue until we have a few chunks (context should be set by "finish" event)
-                if len(chunks) >= 3:
-                    break
-        finally:
-            # Properly close the generator to avoid the warning
-            await generator.aclose()
-
-        # Verify that _current_reasoning_context is set (happens on "finish" event)
-        # Note: In the new refactored design, we need to consume more events to reach "finish"
-        if hasattr(mock_reasoning_agent, '_current_reasoning_context'):
-            assert mock_reasoning_agent._current_reasoning_context is not None
-            assert "steps" in mock_reasoning_agent._current_reasoning_context
-            assert "tool_results" in mock_reasoning_agent._current_reasoning_context
-        else:
-            # The test broke before the "finish" event - this is expected behavior
-            # The context is only set when we reach the end of reasoning
-            assert True  # Test documents that context setting happens at the end
-
-
-class TestStreamingVsNonStreamingConsistency:
-    """Integration tests to ensure streaming and non-streaming give consistent results."""
-
-    @pytest.mark.integration
-    @pytest.mark.skipif(os.getenv("SKIP_CI_TESTS") == "true", reason="Skipped in CI")
-    @pytest.mark.asyncio
-    async def test_streaming_and_non_streaming_both_include_tool_data(self):
-        """
-        Integration test: Both streaming and non-streaming should include actual tool data
-        when tools execute successfully.
-
-        Uses in-process testing with FastAPI TestClient and in-memory MCP servers.
-        """
-        # Skip if no OpenAI key (this is an integration test)
-        if not os.getenv("OPENAI_API_KEY"):
-            pytest.skip("OPENAI_API_KEY not set")
-
-        # Create reasoning agent with in-memory MCP server for testing
-        async def create_test_reasoning_agent() -> ReasoningAgent:
-            # Create MCP manager with in-memory server
-            # Create in-memory MCP server using FastMCP
-
-            server = FastMCP("test_server")
-
-            @server.tool
-            def weather_api(location: str) -> dict:
-                """Get weather information for a location."""
-                return {"location": location, "temperature": "25°C", "condition": "Partly cloudy"}
-
-            # Convert to tools
-            client = Client(server)
-            tools = await to_tools(client)
-
-            # Create and initialize prompt manager
-            prompt_manager = PromptManager()
-            await prompt_manager.initialize()
-
-            return ReasoningAgent(
-                base_url="https://api.openai.com/v1",
-                api_key=os.getenv("OPENAI_API_KEY"),
-                http_client=httpx.AsyncClient(),
-                tools=tools,
-                prompt_manager=prompt_manager,
-            )
-
-        # Create the test agent
-        test_agent = await create_test_reasoning_agent()
-
-        # Override dependencies to use our test setup
-        app.dependency_overrides[get_reasoning_agent] = lambda: test_agent
-        # No MCP manager override needed with new tool architecture
-        app.dependency_overrides[get_prompt_manager] = lambda: test_agent.prompt_manager
-
-        try:
-            with TestClient(app) as client:
-                # Test non-streaming request
-                non_streaming_response = client.post(
-                    "/v1/chat/completions",
-                    json={
-                        "model": "gpt-4o-mini",
-                        "messages": [{"role": "user", "content": "What's the weather in Tokyo? Use the weather_api tool."}],  # noqa: E501
-                        "stream": False,
-                    },
-                )
-
-                assert non_streaming_response.status_code == 200
-                non_streaming_data = non_streaming_response.json()
-                non_streaming_content = non_streaming_data["choices"][0]["message"]["content"]
-
-                # Test streaming request
-                streaming_response = client.post(
-                    "/v1/chat/completions",
-                    json={
-                        "model": "gpt-4o-mini",
-                        "messages": [{"role": "user", "content": "What's the weather in Tokyo? Use the weather_api tool."}],  # noqa: E501
-                        "stream": True,
-                    },
-                )
-
-                assert streaming_response.status_code == 200
-
-                # Parse streaming response
-                streaming_content = ""
-                for line in streaming_response.iter_lines():
-                    if line.startswith("data: ") and not line.startswith("data: [DONE]"):
-                        try:
-                            chunk_data = json.loads(line[6:])
-                            if chunk_data.get("choices") and chunk_data["choices"][0].get("delta", {}).get("content"):  # noqa: E501
-                                streaming_content += chunk_data["choices"][0]["delta"]["content"]
-                        except json.JSONDecodeError:
-                            continue
-
-                # Both should contain actual weather data (temperature, condition, etc.)
-                # They should NOT contain "tool did not execute successfully" type messages
-                weather_indicators = [
-                    "temperature", "°C", "°F", "sunny", "cloudy", "rain", "condition", "tokyo",
-                ]
-                failure_indicators = ["did not execute", "failed", "unavailable", "error"]
-
-                # Check non-streaming
-                has_weather_data_non_streaming = any(
-                    indicator.lower() in non_streaming_content.lower()
-                    for indicator in weather_indicators
-                )
-                has_failure_non_streaming = any(
-                    indicator.lower() in non_streaming_content.lower()
-                    for indicator in failure_indicators
-                )
-
-                # Check streaming
-                has_weather_data_streaming = any(
-                    indicator.lower() in streaming_content.lower()
-                    for indicator in weather_indicators
-                )
-                has_failure_streaming = any(
-                    indicator.lower() in streaming_content.lower()
-                    for indicator in failure_indicators
-                )
-
-                # Both should have weather data and no failure messages
-                assert has_weather_data_non_streaming, (
-                    f"Non-streaming missing weather data: {non_streaming_content[:200]}"
-                )
-                assert has_weather_data_streaming, (
-                    f"Streaming missing weather data: {streaming_content[:200]}"
-                )
-                assert not has_failure_non_streaming, (
-                    f"Non-streaming has failure message: {non_streaming_content[:200]}"
-                )
-                assert not has_failure_streaming, (
-                    f"Streaming has failure message: {streaming_content[:200]}"
-                )
-
-        finally:
-            # Clean up dependency override
-            app.dependency_overrides.clear()
-
-    @pytest.mark.integration
-    @pytest.mark.skipif(os.getenv("SKIP_CI_TESTS") == "true", reason="Skipped in CI")
-    @pytest.mark.asyncio
-    async def test_tool_arguments_and_results_in_streaming_events(self):
-        """
-        Integration test: Verify that tool arguments and results are properly
-        included in streaming reasoning events.
-        """
-        # Skip if no OpenAI key (this is an integration test)
-        if not os.getenv("OPENAI_API_KEY"):
-            pytest.skip("OPENAI_API_KEY not set")
-
-        # Create reasoning agent with in-memory MCP server for testing
-        async def create_test_reasoning_agent() -> ReasoningAgent:
-            # Create MCP manager with in-memory server
-            # Create in-memory MCP server using FastMCP
-
-            server = FastMCP("test_server")
-
-            @server.tool
-            def weather_api(location: str) -> dict:
-                """Get weather information for a location."""
-                return {"location": location, "temperature": "25°C", "condition": "Partly cloudy"}
-
-            # Convert to tools
-            client = Client(server)
-            tools = await to_tools(client)
-
-            # Create and initialize prompt manager
-            prompt_manager = PromptManager()
-            await prompt_manager.initialize()
-
-            return ReasoningAgent(
-                base_url="https://api.openai.com/v1",
-                api_key=os.getenv("OPENAI_API_KEY"),
-                http_client=httpx.AsyncClient(),
-                tools=tools,
-                prompt_manager=prompt_manager,
-            )
-
-        # Create the test agent
-        test_agent = await create_test_reasoning_agent()
-
-        # Override dependencies to use our test setup
-        app.dependency_overrides[get_reasoning_agent] = lambda: test_agent
-        # No MCP manager override needed with new tool architecture
-        app.dependency_overrides[get_prompt_manager] = lambda: test_agent.prompt_manager
-
-        try:
-            with TestClient(app) as client:
-                # Test streaming request to capture tool events
-                streaming_response = client.post(
-                    "/v1/chat/completions",
-                    json={
-                        "model": "gpt-4o-mini",
-                        "messages": [{"role": "user", "content": "Get the weather for Tokyo using the weather_api tool."}],  # noqa: E501
-                        "stream": True,
-                    },
-                )
-
-                assert streaming_response.status_code == 200
-
-                # Parse streaming response and collect tool events
-                tool_start_events = []
-                tool_complete_events = []
-
-                for line in streaming_response.iter_lines():
-                    if line.startswith("data: ") and not line.startswith("data: [DONE]"):
-                        try:
-                            chunk_data = json.loads(line[6:])
-                            if chunk_data.get("choices") and chunk_data["choices"][0].get("delta", {}).get("reasoning_event"):  # noqa: E501
-                                event = chunk_data["choices"][0]["delta"]["reasoning_event"]
-
-                                if event.get("type") == "tool_execution":
-                                    if event.get("status") == "in_progress":
-                                        tool_start_events.append(event)
-                                    elif event.get("status") == "completed":
-                                        tool_complete_events.append(event)
-                        except json.JSONDecodeError:
-                            continue
-
-                # Verify tool start events contain arguments
-                assert len(tool_start_events) > 0, "No tool start events found"
-                start_event = tool_start_events[0]
-
-                assert "metadata" in start_event, "Tool start event missing metadata"
-                assert "tool_predictions" in start_event["metadata"], "Tool start event missing tool_predictions in metadata"  # noqa: E501
-
-                tool_predictions = start_event["metadata"]["tool_predictions"]
-                assert len(tool_predictions) > 0, "No tool predictions in start event"
-
-                # Check that tool predictions have arguments
-                prediction = tool_predictions[0]
-                assert "arguments" in prediction or hasattr(prediction, "arguments"), "Tool prediction missing arguments"  # noqa: E501
-
-                # Verify tool complete events contain results
-                assert len(tool_complete_events) > 0, "No tool complete events found"
-                complete_event = tool_complete_events[0]
-
-                assert "metadata" in complete_event, "Tool complete event missing metadata"
-                assert "tool_results" in complete_event["metadata"], "Tool complete event missing tool_results in metadata"  # noqa: E501
-
-                tool_results = complete_event["metadata"]["tool_results"]
-                assert len(tool_results) > 0, "No tool results in complete event"
-
-                # Check that tool results have actual result data
-                result = tool_results[0]
-                assert hasattr(result, "result") or "result" in result, "Tool result missing result data"  # noqa: E501
-                assert hasattr(result, "tool_name") or "tool_name" in result, "Tool result missing tool_name"  # noqa: E501
-
-        finally:
-            # Clean up dependency override
-            app.dependency_overrides.clear()
